pyproject.toml
[tool.poetry]
name = "magisterka"
version = "0.1.0"
description = ""
authors = ["Wajdzik <Zogfryt@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.6"
numpy = "<2.0"
neo4j = "^5.27.0"
streamlit = "^1.41.1"
spacy-download = "^1.1.0"
plotly = "^5.24.1"
python-dotenv = "^1.0.1"
torch = "^2.5.1"
spacy = {extras = ["cuda12x"], version = "^3.8.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
